# chapter13/whatstore/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Whatstore API",
        "version": "0.1"
    },
    "paths": {
...
        "/storeapi/orderapi/{key}": {
            "post": {
                "operationId": "21e26631c3813779ca92697a7cab4426",
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "description": "order code",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "insert a purchase order"
                    }
                }
            }
        },
        "/storeapi/productbasketapi/{key}": {
            "get": {
                "operationId": "8d134b4fb824c9c283edd9497d4148d9",
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "description": "product code",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "get a product"
                    }
                }
            },
            "put": {
                "operationId": "d5d3cbc224640b469c39dd5bec69cb03",
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "description": "product code",
                        "required": false
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "code",
                                    "name",
                                    "price",
                                    "amount"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "float"
                                    },
                                    "amount": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "update a product"
                    }
                }
            },
            "post": {
                "operationId": "87f8c8024bac79df3bed7a6f24d545b9",
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "description": "product code",
                        "required": false
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "code",
                                    "name",
                                    "price",
                                    "amount"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "float"
                                    },
                                    "amount": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "insert a product"
                    }
                }
            },
            "delete": {
                "operationId": "12cd666347192139d422eee20261443e",
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "description": "product code",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "remove a product"
                    }
                }
            }
        }
    }
}